Molecule rendering, part 1: SMILES parser
Summary:
For the hackathon, I made some perseus widgets that can render simple 2D
molecule structure drawings into perseus content.

The general approach was to start with a text-based molecule representation
called SMILES, which is easy to look up for essentially any molecule of
interest. This format is then parsed into an intermediate tree representation
in the smiles-parser module. The tree is then converted to a list of rendering
instructions, which say what to draw but not how. (e.g. "draw a line from [x,y]
to [w,z]", but in the form of a javascript object, rather than a human-readable
string) This happens in the module molecule-layout.js. Finally, the rendering
instructions are actually drawn into an html5 canvas 2D drawing context by the
molecule-drawing.js module.

This revision is the first part of this rendering pipeline: the SMILES parser.
This takes in the SMILES string (e.g. "CC(=O)C"), and emits a parse tree
representing the molecule.

Test Plan: `make test`

Reviewers: alex

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D21679